Fix passive touch event handling in chrome
Google in their infinite wisdom have found a way
of speeding up touch event handling, by adding the
option to mark event handlers as ‘passive’, which
is a way to tell chrome that the event handler
doesn’t need to call `event.preventDefault()`, and
that Chrome doesn’t need to worry about scroll events
being prevented in event handlers.

This apparently allows chrome to be quicker. See this
article for a fuller explanation:
https://developers.google.com/web/updates/2017/01/scrolling-intervention

From chrome 56 onwards, all touch event handlers are
treated as passive. The result is that the preventDefaults
are ignored by chrome and scrolling was happening
at the same time as dragging and dropping making
drag and drop operations on touch devices impossible.

This commit bumps the touch-dnd-custom-events version number.